+++ pygmo/docstrings.cpp
@@ -2660,7 +2660,7 @@ Args:
 
 Raises:
     OverflowError: if *gen* or *seed* is negative or greater than an implementation-defined value
-    ValueError: if *omega* is not in the [0,1] interval, if *eta1*, *eta2* are not in the [0,1] interval, if *max_vel* is not in ]0,1]
+    ValueError: if *omega* is not in the [0,1] interval, if *eta1*, *eta2* are not in the [0,4] interval, if *max_vel* is not in ]0,1]
     ValueError: *variant* is not one of 1 .. 6, if *neighb_type* is not one of 1 .. 4 or if *neighb_param* is zero
 
 The following variants can be selected via the *variant* parameter:
@@ -3935,10 +3935,6 @@ the following method:
 The ``run_evolve()`` method of the UDI will use the input :class:`~pygmo.algorithm`'s
 :func:`~pygmo.algorithm.evolve()` method to evolve the input :class:`~pygmo.population` and, once the evolution
 is finished, it will return the algorithm used for the evolution and the evolved :class:`~pygmo.population`.
-Note that, since internally the :class:`~pygmo.island` class uses a separate thread of execution to provide asynchronous
-behaviour, a UDI needs to guarantee a certain degree of thread-safety: it must be possible to interact with the UDI
-while evolution is ongoing (e.g., it must be possible to copy the UDI while evolution is undergoing, or call the ``get_name()``,
-``get_extra_info()`` methods, etc.), otherwise the behaviour will be undefined.
 
 In addition to the mandatory ``run_evolve()`` method, a UDI may implement the following optional methods:
 
@@ -3951,6 +3947,15 @@ In addition to the mandatory ``run_evolve()`` method, a UDI may implement the fo
 
 See the documentation of the corresponding methods in this class for details on how the optional
 methods in the UDI are used by :class:`~pygmo.island`.
+
+Note that, due to the asynchronous nature of :class:`~pygmo.island`, a UDI has certain requirements regarding
+thread safety. Specifically, ``run_evolve()`` is always called in a separate thread of execution, and consequently:
+
+* multiple UDI objects may be calling their own ``run_evolve()`` method concurrently,
+* in a specific UDI object, any method from the public API of the UDI may be called while ``run_evolve()`` is
+  running concurrently in another thread. Thus, UDI writers must ensure that actions such as copying
+  the UDI, calling the optional methods (such as ``get_name()``), etc. can be safely performed while the island
+  is evolving.
 
 An island can be initialised in a variety of ways using keyword arguments:
 
@@ -4404,7 +4409,7 @@ and :attr:`~pygmo.nlopt.replacement` attributes.
 
 .. seealso::
 
-   The `NLopt website <http://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/>`__ contains a detailed description
+   The `NLopt website <https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/>`__ contains a detailed description
    of each supported solver.
 
 This constructor will initialise an :class:`~pygmo.nlopt` object which will use the NLopt algorithm specified by
@@ -4442,7 +4447,7 @@ See also the docs of the C++ class :cpp:class:`pagmo::nlopt`.
 
 .. seealso::
 
-   The `NLopt website <http://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/>`__ contains a detailed
+   The `NLopt website <https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/>`__ contains a detailed
    description of each supported solver.
 
 Args: